safeml/requirements.txt
# platform
mysql-connector-python
flask
pytz==2021.1

# SafeML libraries
ultralytics==8.0.149
IPython
scikit-image==0.21.0
scikit-learn==1.3.1

# deepknowledge libraries
tensorflow-datasets
imagecorruptions
pandas==2.0.3